Moved create_subscription to subscription.py
